+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 - Fix: mCaptchaを使用していてもbotプロテクションに関する警告が消えないのを修正
 
 ### Server
+- Enhance: 凍結されたアカウントのフォローリクエストを表示しないように
 - Fix: WSの`readAllNotifications` メッセージが `body` を持たない場合に動作しない問題 #14374
   - 通知ページや通知カラム(デッキ)を開いている状態において、新たに発生した通知が既読されない問題が修正されます。
   - これにより、プッシュ通知が有効な同条件下の環境において、プッシュ通知が常に発生してしまう問題も修正されます。
